Strip version from service catalog endpoint
This client includes the API version in the URL directly where the
former practice was to include it in the service catalog endpoint.
This change removes the version from the last component of the
SC endpoint (if present) for transition purposes.

Note that this does not generalize to the other APIs
where the version is not the last component of the SC endpoint.

Change-Id: Ie04c38d80b17a171482e195aa1c633b6b6974042