Add support for passing image size to Glance API
Introduced a new command-line option --size for the image-create and
image-create-via-import commands. Added the `x-openstack-image-size` header
to transmit the image size from image-upload and image-stage commands
to the Glance API.

If --size is not specified but --file is provided, the script calculates
the file size and includes it in the `x-openstack-image-size` header when
invoking image-upload and image-stage commands.

Change-Id: I7572f8a5d42a9968b940ed71eecbe3028e92877e
Signed-off-by: Abhishek Kekane <akekane@redhat.com>